add 	conditional jump

+++ reboot/test-rb.bash
@@ -4,7 +4,14 @@ DATE=`date '+%F %R'`
 F_DATE=`date '+%Y%m%d'`
 FILE="/var/log/load/reboot_dir/test-rb_${F_DATE}.log"
 
-CLK=`cat /sys/devices/system/clocksource/clocksource0/current_clocksource`
-echo "[$DATE]: $CLK" >> $FILE
+num_line=`wc -l $FILE`
 
-reboot
+if [ $num_line -lt 100 ]
+then
+    CLK=`cat /sys/devices/system/clocksource/clocksource0/current_clocksource`
+    echo "[$DATE]: $CLK" >> $FILE
+
+    reboot
+else
+    echo "already rebooted 100 times. finished this script." >> $FILE
+fi
